a lot of stuff, check the description
fixed an issue with all factories when deleting individual items.
Created AllegrWindow, a way to wrap allegro display.
Now the equal operator was moved to WrittenBox.
Now you can initiate a writtenBox without the need to create a separate font for it.
Now When you press a button, it changes color to reflect that is pressed, the only caveat is that you have to manually unpress it.
New getters for writtenBox
added a way to clear all text in writable box.